% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_title}
\alias{convert_title}
\title{Replace R markdown header with ## title}
\usage{
convert_title(md_file)
}
\arguments{
\item{md_file}{string; markdown file text}
}
\description{
Replace R markdown header with title only
}
