% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_image_numbers}
\alias{generate_image_numbers}
\title{Generate image numbers}
\usage{
generate_image_numbers(imgs)
}
\arguments{
\item{imgs}{character vector of png files to be converted to Whitehall format
(!!n). The filename must start and end with a number and have text in
between (eg, "1 abcd 1.png")}
}
\description{
Converts image file names to a dataframe, with a field containing the
original image name and one for the Whitehall publishing software
}
