% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_md}
\alias{convert_md}
\title{Convert markdown file to Whitehall publishing format}
\usage{
convert_md(path, images_folder = "images")
}
\arguments{
\item{path}{string; filename (including path) to *.md file for conversion}

\item{images_folder}{string; folder containing images for *.md file. Defaults
to "images"}
}
\description{
Convert markdown file to Whitehall (gov.uk) publishing format
}
